fixed bug by uncommenting ClearingDecoy functions
git-svn-id: https://www.concebo.eu/svn/paola/KIR_code@23 45e2b9c2-40ca-4359-bf5c-84bf3056f186

+++ src/Genes.h
@@ -54,7 +54,7 @@ protected:
 	int geneID;
 	int specificity;
 	bool pseudoGene;
-	int mID;
+	unsigned long int mID;
 
 
 };
